clang-tidy

+++ tasks/ilin_a_gaussian_method_horizontal_band_scheme/tests/functional/main.cpp
@@ -7,7 +7,6 @@
 #include <random>
 #include <string>
 #include <tuple>
-#include <utility>
 #include <vector>
 
 #include "ilin_a_gaussian_method_horizontal_band_scheme/common/include/common.hpp"
@@ -63,7 +62,7 @@ class IlinARunFuncTestsProcesses : public ppc::util::BaseRunFuncTests<InType, Ou
     int band_width = std::max(1, size / 4);
     band_width = std::min(band_width, size);
 
-    std::vector<double> matrix(static_cast<size_t>(size) * band_width, 0.0);
+    std::vector<double> matrix(static_cast<size_t>(size) * static_cast<size_t>(band_width), 0.0);
     std::vector<double> vector(static_cast<size_t>(size), 0.0);
     expected_solution_.resize(static_cast<size_t>(size));
 
@@ -83,14 +82,14 @@ class IlinARunFuncTestsProcesses : public ppc::util::BaseRunFuncTests<InType, Ou
           int band_idx = (i - j + band_width - 1);
           if (band_idx >= 0 && band_idx < band_width) {
             double val = dist(gen) * 0.1;
-            matrix[static_cast<size_t>(i) * band_width + band_idx] = val;
+            matrix[static_cast<size_t>(i) * static_cast<size_t>(band_width) + band_idx] = val;
             diag_sum += std::fabs(val);
           }
         }
       }
 
       int diag_band_idx = band_width - 1;
-      matrix[static_cast<size_t>(i) * band_width + diag_band_idx] = diag_sum + dist(gen) + 10.0;
+      matrix[static_cast<size_t>(i) * static_cast<size_t>(band_width) + diag_band_idx] = diag_sum + dist(gen) + 10.0;
     }
 
     for (int i = 0; i < size; ++i) {
@@ -99,7 +98,8 @@ class IlinARunFuncTestsProcesses : public ppc::util::BaseRunFuncTests<InType, Ou
         if (j <= i) {
           int band_idx = (i - j + band_width - 1);
           if (band_idx >= 0 && band_idx < band_width) {
-            sum += matrix[static_cast<size_t>(i) * band_width + band_idx] * expected_solution_[static_cast<size_t>(j)];
+            sum += matrix[static_cast<size_t>(i) * static_cast<size_t>(band_width) + band_idx] *
+                   expected_solution_[static_cast<size_t>(j)];
           }
         }
       }
